Extract trunk methods
